Move angus client sdk in "angus.client" subpackage. The package angus is now a namespace, angus.connect is deprecated. Please use angus.client.connect instead.
