+++ angus/__init__.py
@@ -18,135 +18,14 @@
 # under the License.
 
 
-import argparse
-import json
 import logging
-import os
 import pkgutil
-import sys
-
-import requests
-
-import angus.cloud
-import angus.rest
-import angus.version
-
-
-__path__ = pkgutil.extend_path(__path__, __name__)
-
-__version__ = angus.version.__version__
-__updated__ = "2017-08-08"
-__author__ = "Aurélien Moreau"
-__copyright__ = "Copyright 2015-2017, Angus.ai"
-__credits__ = ["Aurélien Moreau", "Gwennael Gate"]
-__license__ = "Apache v2.0"
-__maintainer__ = "Aurélien Moreau"
-__status__ = "Production"
+import client
 
 LOGGER = logging.getLogger('AngusSDK')
 
-requests.packages.urllib3.disable_warnings()
+__path__ = pkgutil.extend_path(__path__, __name__)
 
-def parse_cmd_line(argv):
-    """Parse command line argument. See -h option.
-    """
-    formattter_class = argparse.RawDescriptionHelpFormatter
-
-    parser = argparse.ArgumentParser(description='Angus SDK',
-                                     formatter_class=formattter_class)
-
-    parser.add_argument('-c', '--configuration', metavar='filename', type=str,
-                        help='configuration file.')
-
-    parser.add_argument('-id', '--clientid', metavar='clientid', type=str,
-                        help='the client id.')
-
-    parser.add_argument('-a', '--access', metavar='access', type=str,
-                        help='the access token.')
-
-    parser.add_argument('-v', '--verbose', dest='vlevel', action='count',
-                        default=0,
-                        help='increases log vervosity for each occurence.')
-
-    parser.add_argument('-ca', '--capath', metavar='capath', type=str,
-                        help='path to a trusted ca bundle.')
-
-    parser.add_argument('-r', '--root', metavar='root', type=str,
-                        help='the angus gate url')
-
-    parser.add_argument('--version', action='version',
-                        version="%(prog)s {}".format(__version__))
-
-    args = parser.parse_args(argv)
-
-    return args
-
-
-def get_default_configuration(argv=""):
-    """Get and parse the default configuration in ~/.angusdk/config.json
-    """
-    args = parse_cmd_line(argv)
-
-    # Set the logger level
-    logging.basicConfig()
-    LOGGER.setLevel(max(3 - args.vlevel, 0) * 10)
-
-    # Get the configuration file
-    default_file = os.path.realpath("./config.json")
-    if args.configuration is not None:
-        if os.path.isfile(args.configuration):
-            conf_file = args.configuration
-        else:
-            LOGGER.error("The configuration file '%s' does not exist",
-                         args.configuration)
-            sys.exit(-1)
-    elif os.path.isfile(default_file):
-        conf_file = default_file
-    else:
-        default_file = os.path.expanduser("~/.angusdk/config.json")
-
-        if os.path.isfile(default_file):
-            conf_file = default_file
-        else:
-            conf_file = None
-
-    conf = angus.rest.Configuration()
-
-    # Apply the configuration file if it exists
-    if conf_file is not None:
-        with open(conf_file, 'r') as fconf:
-            conf_data = json.loads(fconf.read())
-            if 'client_id'in conf_data and 'access_token' in conf_data:
-                conf.set_credential(
-                    conf_data['client_id'],
-                    conf_data['access_token'])
-
-            if 'ca_path' in conf_data:
-                conf.set_ca_path(conf_data['ca_path'])
-
-            if 'default_root' in conf_data:
-                conf.default_root = conf_data['default_root']
-
-    if args.clientid and args.access:
-        conf.set_credential(args.clientid, args.access)
-
-    if args.root:
-        conf.default_root = args.root
-
-    if args.capath:
-        conf.set_ca_path(args.capath)
-
-    return conf
-
-
-def connect(url=None, conf=None, client_id=None,
-            access_token=None, verify=True):
-    """Create a resource root with the given configuration.
-    """
-    if None not in [client_id, access_token]:
-        conf = angus.rest.Configuration()
-        conf.set_credential(client_id, access_token)
-        if not verify:
-            conf.do_not_verify()
-
-    return angus.cloud.Root(url, conf)
+def connect(*args, **kwargs):
+    LOGGER.warning("Call to deprecated function angus.connect. Please use angus.client.connect")
+    return client.connect(*args, **kwargs)
