+++ tests/test_stream.py
@@ -21,9 +21,9 @@ import datetime
 import pytest
 import pytz
 
-import angus
+import angus.client
 
-__updated__ = "2017-01-03"
+__updated__ = "2017-08-07"
 __author__ = "Aurélien Moreau"
 __copyright__ = "Copyright 2015-2017, Angus.ai"
 __credits__ = ["Aurélien Moreau", "Gwennael Gate"]
@@ -50,7 +50,7 @@ def service_name(request):
 
 @pytest.fixture(scope="module")
 def root(server, client, token, verify):
-    return angus.connect(
+    return angus.client.connect(
         url=server, client_id=client, access_token=token, verify=verify)
 
 @pytest.fixture(scope="module")
